Initial commit
With this commit the players strategy is to play a brick if the number
of lifes are greater than 1. Otherwise they trash a brick. There is a
method to customize the deal, meaning that you can deal some specific
bricks to the players and then deal the rest as usual.