Saving clues in each brick and use that information
Strategy:
 (1) If the player sees a playable brick and there are clues left then
 they give a clue about that brick.
 (2) If no clues are available or no playable bricks in sight the the
 player checks if they have any playable bricks
 (3) As a last option they trash a brick.

There is one bug that if a player gets the same clue again then the clue
value becomes an invalid value (i.e a value that does not correspondes
to a number or color).

Also the code needs to be more structured so that it is not possible to
cheat.