Add Player.CalculateBrickPlayability and print statistics
- The Player class now contains a method that calculates the probability
that a certain brick can be played based on what the player can see and
which clues the player got about the brick
- Add a method to print a bar diagram showing how many times the team
got certain points.
- The brick class now has a property which indicates where the brick is
(a players hand, trash pile, draw pile or table)
- The player class contains methods to get visible and unvisible bricks.
- The best strategy is updated so a player plays a brick if it is 100%
probability that the player can play it. The player also priorities to
give number clues before color clues. This makes it easier to place the
first one.