Brick importance and refactoring
- Add brick importance method which is useful when determine which
bricks to give clues. I think this gave an additional 0.1 score in
average
- Separate strategy steps into three classes, clue, play and brick. This
makes it easier to build new streategies
- Implemented a strategy where you point at two bricks and the newest
brick is playable. Unfortunately this did not give any better results

Next idea is to remember which clues has been given so that a player
doent give a player about a brick which has already been clued somewhere
else. For instace
 - if player 1 give player 3 a clue about a green 4
 - then player 2 should not give player 1 a clue about a green 4

Also it might be good to give clues about multiple ones in the begining.